# simplechat/artifacts/open_api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Chat APIs
  version: 0.215.37
  description: API for the Simple Chat, handling chats, conversations, documents, user settings, groups, feedback, prompts, models, safety, and admin functions.
servers:
  - url: /api
    description: API Root
...
paths:
...
  /get_file_content:
    post:
      tags:
        - Documents
      summary: Get content of a file previously uploaded within a conversation
      description: Retrieves the extracted content of a file message stored within a specific conversation. Requires `enable_user_workspace` setting to be true and user authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - conversation_id
                - file_id
              properties:
                conversation_id:
                  type: string
                  format: uuid
                  description: The ID of the conversation containing the file.
                file_id:
                  type: string
                  description: The ID of the file message (e.g., 'convid_file_timestamp_rand').
      responses:
        '200':
          description: File content retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_content:
                    type: string
                    description: The combined text content of the file.
                  filename:
                    type: string
                    description: The original name of the file.
                  is_table:
                    type: boolean
                    description: Indicates if the content represents tabular data.
        '400':
          description: Bad Request (Missing conversation_id or file_id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Conversation or File not found in the specified conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database query failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/upload:
    post:
      tags:
        - Documents
      summary: Upload document(s) to the user's workspace
      description: Accepts one or more file uploads, saves them temporarily, creates metadata records in the database with 'Queued' status, and queues background tasks for processing and chunking. Returns immediately with document IDs. Requires `enable_user_workspace` setting to be true and user authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file # Technically the key used by getlist('file')
              properties:
                file: # Although named 'file', this can represent multiple files
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more document files to upload.
      responses:
        '200':
          description: File(s) accepted for background processing. All files processed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '207':
          description: Multi-Status. Some files processed, but errors occurred with others. Check 'errors' array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          description: Bad Request (No file part, no files selected, all files invalid type or failed initial save/queue). Check 'errors' array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse' # May still contain errors array
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Unexpected error during queuing)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Or DocumentUploadResponse with a generic error
  /documents:
    get:
      tags:
        - Documents
      summary: Get user's documents with pagination and filtering
      description: Retrieves a paginated list of documents uploaded by the authenticated user, with filtering options. Requires `enable_user_workspace` setting to be true and user authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: The page number to retrieve.
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            # maximum: 100 # Consider adding a max limit
          description: The number of documents per page.
        - name: search
          in: query
          schema:
            type: string
          description: Search term for file name or title (case-insensitive).
        - name: classification
          in: query
          schema:
            type: string
          description: Filter by document classification category (case-sensitive). Use 'none' to find unclassified.
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author (case-insensitive, exact match in authors array).
        - name: keywords
          in: query
          schema:
            type: string
          description: Filter by keyword (case-insensitive, exact match in keywords array).
        - name: abstract
          in: query
          schema:
            type: string
          description: Search within the document abstract (case-insensitive).
      responses:
        '200':
          description: List of documents retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/DocumentMetadata'
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_count:
                    type: integer
                  needs_legacy_update_check:
                    type: boolean
                    description: Indicates if there are documents needing the legacy update process.
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database query failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /documents/upgrade_legacy:
    post:
      tags:
        - Documents
        - Admin # Often an admin-triggered action
      summary: Upgrade legacy user documents
      description: Triggers a process to update user documents created before a schema change (e.g., adding `percentage_complete`). Requires `enable_user_workspace` and user authentication.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Legacy document upgrade process completed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Upgraded 5 document(s) to the new format."
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Upgrade process failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --- Feedback Routes (route_backend_feedback.py) ---
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error that occurred.
...
    DocumentUploadResponse:
      type: object
      description: Response after attempting to upload one or more documents.
      properties:
        message:
          type: string
          description: Overall status message.
          example: "Processed 2 file(s). Check status periodically."
        document_ids:
          type: array
          items:
            type: string
            format: uuid
          description: List of IDs assigned to the document records successfully queued for processing.
        processed_filenames:
          type: array
          items:
            type: string
          description: List of filenames corresponding to the successfully queued documents.
        errors:
          type: array
          items:
            type: string
          description: List of error messages for files that failed to upload or queue.
...
  securitySchemes:
    bearerAuth: # Defines how API clients authenticate
      type: http
      scheme: bearer
      bearerFormat: JWT # Indicates the type of token expected
      description: "Azure AD issued JWT Bearer token required for authenticated endpoints. Include in the request header as 'Authorization: Bearer <YOUR_TOKEN>'."